Separates edge-cases testing
